+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+
+- add variables to template text (t:)
+- fix template text-spaces
+
+
 ## 1.0.0
 
 - add core logic